Extract SwipePositionHandler for readability
