+++ app/src/main/java/com/example/kdotdi/app/view/cv/SwipeToDeletePositionCallback.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
 
-abstract class SwipeToDeletePositionCallback(context: Context) :
+abstract class SwipeToDeletePositionCallback() :
     ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT) {
     override fun onMove(
         recyclerView: RecyclerView,
